backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tortoise-orm[asyncpg]==0.20.0
aerich==0.7.2
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.6
websockets==12.0
boto3>=1.34.72
python-dotenv==1.0.0
langchain>=1.0.0a13
langgraph>=1.0.0a4
